+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 - Public: Merged `dist/onfido.vendors~crossDevice.min.js` into `dist/onfido.crossDevice.min.js`
 - Public: Merged `dist/onfidoAuth.6.min.js` to `onfidoAuth.min.js`
 - Public: Rearranged `dependencies` & `devDependencies` in `package.json` to reflect correct usage
+- Public: Added a cookie with anonymous uuid for analytics purposes
 - Internal: Upgrade `minimist` to `v1.2.6` to fix vulnerability
 - Public: Added a "microphone & camera" permission screen
 - Internal: Handle errors by callback instead of `throw` for `requestChallenges` and `postToBackend`